Added first draft of dissertation
This dissertation will be updated over the next two weeks. All updates will be reflected on this github page, so make sure you have the most recent version whenever you choose to read it. List of areas under construction will be kept up to date in the ReadMe.md file.